+++ Keyboards/LanguageKeyboards/Swedish/SwedishKeyboardViewController.swift
@@ -1,0 +1,5 @@
+//
+//  SwedishKeyboardViewController.swift
+//
+
+class SwedishKeyboardViewController: KeyboardViewController {}